requirements.txt
fastapi
mangum
pydantic
psycopg2-binary
python-dotenv
boto3
werkzeug